+++ R/rgexf-package.R
@@ -32,10 +32,8 @@ gexf_version <- function(vers="1.3") {
   }
 }
 
-#' @importFrom XML xmlTreeParse xmlNode addChildren asXMLNode saveXML newXMLDoc
-#'  newXMLNode newXMLNamespace xmlAttrs parseXMLAndAdd
-#' @importFrom igraph get.data.frame list.vertex.attributes list.edge.attributes
-#'   E is.directed V set.vertex.attribute set.edge.attribute
+#' @importFrom XML xmlTreeParse xmlNode addChildren asXMLNode saveXML newXMLDoc newXMLNode newXMLNamespace xmlAttrs parseXMLAndAdd
+#' @importFrom igraph get.data.frame list.vertex.attributes list.edge.attributes E is.directed V set.vertex.attribute set.edge.attribute
 #' @importFrom grDevices rgb col2rgb colors
 #' @importFrom utils head
 #' @importFrom servr httd
@@ -76,6 +74,8 @@ NULL
 #' @param file String. Output path where to save the GEXF file.
 #' @param replace Logical. If \code{file} exists, \code{TRUE} would replace the
 #' file.
+#' @param nnodes Integer. Maximum number of nodes to print (default 10).
+#' @param nedges Integer. Maximum number of edges to print (default 10).
 #' @param \dots Ignored
 #' @return \item{list("print.gexf")}{ None (invisible \code{NULL}).}
 #' \item{list("summary.gexf")}{ List containing some \code{gexf} object